Adding Loader component to personal values and reorganise files
