+++ src/contexts/session.tsx
@@ -10,9 +10,7 @@ export const SessionContext = createContext<TSessionProvider | null>(null);
 export const SessionDispatch = createContext<TSessionDispatch | null>(null);
 
 export const SessionProvider: React.FC = ({ children }) => {
-  const [sessionId, setSessionId] = useState<string | null>(
-    'e9717610-31c7-4365-970d-ba910f101bda'
-  );
+  const [sessionId, setSessionId] = useState<string | null>(null);
 
   return (
     <SessionContext.Provider value={sessionId}>
