+++ src/pages/PersonalValues.tsx
@@ -2,6 +2,7 @@ import React from 'react';
 import { useHistory } from 'react-router-dom';
 import { Typography, Button, Grid, makeStyles, Box } from '@material-ui/core';
 import { ReactComponent as Logo } from '../assets/cm-logo.svg';
+import Loader from '../components/Loader';
 import ROUTES from '../components/Router/RouteConfig';
 
 import CMCard from '../components/CMCard';
@@ -25,7 +26,7 @@ const PersonalValues: React.FC = () => {
 
   //TODO: fix loading spinner and maybe condition for showing it...
   if (Object.keys(climatePersonality).length < 1) {
-    return <p>loading...</p>;
+    return <Loader />;
   }
   return (
     <Grid
